REF: move config widgets out of bin, split widgets.py into submodules
