+++ atef/widgets/__init__.py
@@ -1,0 +1,2 @@
+from .config import *  # NOQA
+from .procedure import *  # NOQA
